+++ classes/output/core_renderer_toolbox.php
@@ -99,7 +99,8 @@ trait core_renderer_toolbox {
         $usermenuitems[] = ['enableeditprofile', false, $CFG->wwwroot . '/user/edit.php', get_string('editmyprofile'),
             \theme_adaptable\toolbox::getfontawesomemarkup('cog', ['mr-1']), ];
         $usermenuitems[] = ['enableaccesstool', false, $CFG->wwwroot . '/local/accessibilitytool/manage.php',
-            get_string('enableaccesstool', 'theme_adaptable'), \theme_adaptable\toolbox::getfontawesomemarkup('low-vision', ['mr-1']), ];
+            get_string('enableaccesstool', 'theme_adaptable'),
+            \theme_adaptable\toolbox::getfontawesomemarkup('low-vision', ['mr-1']), ];
         $usermenuitems[] = ['enableprivatefiles', false, $CFG->wwwroot . '/user/files.php',
             get_string('privatefiles', 'block_private_files'), \theme_adaptable\toolbox::getfontawesomemarkup('file', ['mr-1']), ];
         if (\theme_adaptable\toolbox::kalturaplugininstalled()) {
@@ -121,7 +122,8 @@ trait core_renderer_toolbox {
         $usermenuitems[] = ['enablefeed', false, $CFG->wwwroot . '/report/myfeedback/index.php',
             get_string('enablefeed', 'theme_adaptable'), \theme_adaptable\toolbox::getfontawesomemarkup('bullhorn', ['mr-1']), ];
         $usermenuitems[] = ['enablecalendar', false, $CFG->wwwroot . '/calendar/view.php',
-            get_string('pluginname', 'block_calendar_month'), \theme_adaptable\toolbox::getfontawesomemarkup('calendar', ['mr-1']), ];
+            get_string('pluginname', 'block_calendar_month'),
+            \theme_adaptable\toolbox::getfontawesomemarkup('calendar', ['mr-1']), ];
 
         $returnurl = $this->page->url->out_as_local_url(false);
         $context = context_course::instance($COURSE->id);
@@ -492,9 +494,8 @@ trait core_renderer_toolbox {
      *       attribute.  This version needed for 'Edit button keep position' in adaptable.js.
      *
      * @param moodle_url $url The URL + params to send through when clicking the button.
-     * @parmm string $method Not used.
+     * @param string $method Not used.
      * @return string HTML the button
-     * Written by G J Barnard
      */
     public function edit_button(moodle_url $url, string $method = 'post') {
         $url->param('sesskey', sesskey());
@@ -745,16 +746,19 @@ trait core_renderer_toolbox {
                             /* Here's the crucial bit.  Check if span number is 0,
                                or $displayall is true (override) and if so, print it out. */
                             if ($spannumber == 0 || $displayall) {
-                                $blockclass = $block['classnamebeginswith'] . chr(96 + $y);
-                                $missingblock = $this->blocks($blockclass, 'block');
+                                $blockregion = $block['classnamebeginswith'] . chr(96 + $y);
+                                $displayregion = $this->page->apply_theme_region_manipulations($blockregion);
 
                                 // Check if the block actually has content to display before displaying.
-                                if (strip_tags($missingblock)) {
+                                if ($this->page->blocks->region_has_content($displayregion, $this)) {
                                     if ($adminediting) {
-                                        $missingblocks .= '<em>ORPHANED BLOCK - Originally displays in: <strong>' .
-                                                get_string('region-' . $blockclass, 'theme_adaptable') . '</strong></em>';
+                                        $missingblocks .= get_string(
+                                            'orphanedblock',
+                                            'theme_adaptable',
+                                            get_string('region-' . $blockregion, 'theme_adaptable')
+                                        );
                                     }
-                                    $missingblocks .= $missingblock;
+                                    $missingblocks .= $this->blocks($blockregion, 'block');
                                 }
                             }
                         }
@@ -1288,7 +1292,8 @@ trait core_renderer_toolbox {
                     $mysitesmaxlengthhidden,
                     $this->page->theme->settings,
                     $this->page,
-                    $this);
+                    $this
+                );
             }
 
             if (!empty($this->page->theme->settings->enablethiscourse)) {
@@ -1296,7 +1301,8 @@ trait core_renderer_toolbox {
                     $branchlabel = '';
                     $branchtitle = get_string('thiscourse', 'theme_adaptable');
                     if ($navbardisplayicons) {
-                        $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('sitemap', ['fa-lg']).'<span class="menutitle">';
+                        $branchlabel .=
+                            \theme_adaptable\toolbox::getfontawesomemarkup('sitemap', ['fa-lg']) . '<span class="menutitle">';
                     }
                     $branchlabel .= $branchtitle;
                     if ($navbardisplayicons) {
@@ -1322,7 +1328,12 @@ trait core_renderer_toolbox {
                     if ($this->page->theme->settings->displayparticipants) {
                         $branchtitle = get_string('people', 'theme_adaptable');
                         $branchlabel = \theme_adaptable\toolbox::getfontawesomemarkup(
-                            'users', ['icon', 'mr-2'], [], '', $branchtitle) . $branchtitle;
+                            'users',
+                            ['icon', 'mr-2'],
+                            [],
+                            '',
+                            $branchtitle
+                        ) . $branchtitle;
                         $branchurl = new moodle_url('/user/index.php', ['id' => $this->page->course->id]);
                         $branch->add($branchlabel, $branchurl, $branchtitle, $branchmenusort);
                     }
@@ -1375,7 +1386,7 @@ trait core_renderer_toolbox {
                                 $branchmenusort
                             );
                         } else {
-                            $icon = $this->pix_icon('monologo', get_string('pluginname', 'mod_'.$modname), $modname);
+                            $icon = $this->pix_icon('monologo', get_string('pluginname', 'mod_' . $modname), $modname);
                             $branchmenusort++;
                             $branch->add(
                                 $icon . $modfullname,
@@ -1466,7 +1477,12 @@ trait core_renderer_toolbox {
         if (!empty($sectionsformnenu)) { // Rare but possible!
             $branchtitle = get_string('sections', 'theme_adaptable');
             $branchlabel = \theme_adaptable\toolbox::getfontawesomemarkup(
-                'list-ol', ['icon', 'fa-lg'], [], '', $branchtitle) . $branchtitle;
+                'list-ol',
+                ['icon', 'fa-lg'],
+                [],
+                '',
+                $branchtitle
+            ) . $branchtitle;
             $branch = $menu->add($branchlabel, null, $branchtitle, 100003);
 
             foreach ($sectionsformnenu as $sectionformenu) {
@@ -1538,7 +1554,8 @@ trait core_renderer_toolbox {
 
         if (!empty($logosetarea)) {
             // Logo.
-            $responsivelogo = (empty($this->page->theme->settings->responsivelogo)) ? '' : ' ' . $this->page->theme->settings->responsivelogo;
+            $responsivelogo = (empty($this->page->theme->settings->responsivelogo)) ? '' : ' ' .
+                $this->page->theme->settings->responsivelogo;
             $logomarkup = '<div class="pb-2 pr-3 pt-2 bd-highlight' . $responsivelogo . '">';
             $logo = '<img src=' . $this->page->theme->setting_file_url($logosetarea, $logosetarea) . ' id="logo"';
             $logo .= ' alt="' . get_string('logo', 'theme_adaptable') . '">';
@@ -1548,7 +1565,8 @@ trait core_renderer_toolbox {
                 $logomarkup .= $logo;
             } else {
                 // Logo is a link to site homepage when there is no navbar.
-                $logomarkup .= '<a href=' . $CFG->wwwroot . ' aria-label="' . get_string('home') . '" title="' . format_string($SITE->fullname) . '">';
+                $logomarkup .= '<a href=' . $CFG->wwwroot . ' aria-label="' . get_string('home') . '" title="' .
+                    format_string($SITE->fullname) . '">';
                 $logomarkup .= $logo;
                 $logomarkup .= '</a>';
             }
@@ -1595,7 +1613,8 @@ trait core_renderer_toolbox {
             $sitetitle = $this->get_site_title();
             if (empty($sitetitle)) {
                 if (!empty($categoryheadercustomtitle)) {
-                    $titlemarkup .= '<div id="headertitle" class="bd-highlight pt-2 ' . $themesettings->responsiveheadertitle . '">';
+                    $titlemarkup .= '<div id="headertitle" class="bd-highlight pt-2 ' .
+                        $themesettings->responsiveheadertitle . '">';
                     $titlemarkup .= '<h1><span id="categorytitle">' . format_string($categoryheadercustomtitle) . '</span></h1>';
                     $titlemarkup .= '</div>';
                 }
@@ -1614,6 +1633,11 @@ trait core_renderer_toolbox {
         return $titlemarkup;
     }
 
+    /**
+     * Get the site title.
+     *
+     * return string Site title.
+     */
     protected function get_site_title() {
         global $SITE;
 
@@ -1642,6 +1666,11 @@ trait core_renderer_toolbox {
         return $sitetitle;
     }
 
+    /**
+     * Get the course title.
+     *
+     * return string Course title.
+     */
     protected function get_course_title() {
         global $COURSE;
 
@@ -1668,7 +1697,7 @@ trait core_renderer_toolbox {
                 (!empty($themesettings->coursetitlemaxwidth) ? $themesettings->coursetitlemaxwidth : 0);
             // Check max width of course title and trim if appropriate.
             if (($coursetitlemaxwidth > 0) && ($coursetitle <> '')) {
-                if (strlen($coursetitle) > $coursetitlemaxwidth) {
+                if (\core_text::strlen($coursetitle) > $coursetitlemaxwidth) {
                     $coursetitle = \core_text::substr($coursetitle, 0, $coursetitlemaxwidth) . " ...";
                 }
             }
@@ -1685,7 +1714,7 @@ trait core_renderer_toolbox {
      *
      * @return string Markup.
      */
-    public function context_header($headerinfo = null, $headinglevel = 1) : string {
+    public function context_header($headerinfo = null, $headinglevel = 1): string {
         $headerinfo = [];
         $headerinfo['heading'] = $this->get_course_title();
         return parent::context_header($headerinfo, $headinglevel);
@@ -2308,7 +2337,7 @@ trait core_renderer_toolbox {
             $themesettings = \theme_adaptable\toolbox::get_settings();
             $retr = $localtoolbox->generate_login($logincontent, $themesettings);
         } else {
-            $retr = new stdClass;
+            $retr = new stdClass();
             $retr->header = false;
             $retr->footer = false;
         }