+++ settings/block_settings.php
@@ -26,7 +26,8 @@
 defined('MOODLE_INTERNAL') || die;
 
 if ($ADMIN->fulltree) {
-    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_blocks', get_string('settingspageblocksettings', 'theme_adaptable'));
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_blocks',
+        get_string('settingspageblocksettings', 'theme_adaptable'));
 
     // General.
     $name = 'theme_adaptable/settingsblocksgeneral';