+++ settings/navbar_tools_menu.php
@@ -26,7 +26,8 @@ defined('MOODLE_INTERNAL') || die;
 
 // Settings for tools menus.
 if ($ADMIN->fulltree) {
-    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_navbar_tools_menu', get_string('toolsmenu', 'theme_adaptable'), true);
+    $page = new \theme_adaptable\admin_settingspage('theme_adaptable_navbar_tools_menu',
+        get_string('toolsmenu', 'theme_adaptable'), true);
 
     $page->add(new admin_setting_heading(
         'theme_adaptable_toolsmenu',